backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
prometheus-client==0.20.0
psycopg2-binary
